REMOVE concurrency
Loading functions/triggers/views concurrently was fine during testing on
a mock database, but it does not play well in a real codebase: as soon
as functions start referencing each others, it causes race conditions
when a function body uses an other function that just has been dropped
and is not recreated yet.

Sequential loading will have to do.